requirements.txt
google-generativeai
speechrecognition
pyttsx3
pyaudio
flask[async]